Factor out spreadsheet and rename datatable to datamanager
